+++ packages/side-effects/side-effect-title/src/index.ts
@@ -1,25 +1,12 @@
-import { Observer, Emitted } from "@curi/router";
+import { Observer, Emitted, Response } from "@curi/router";
 
-export interface TitleOptions {
-  prefix?: string;
-  suffix?: string;
-  delimiter?: string;
-}
+export type TitleBuilder = (response: Response) => string;
 
 export default function createTitleSideEffect(
-  options?: TitleOptions
+  callback: TitleBuilder
 ): Observer {
-  const { prefix = "", suffix = "", delimiter = "" } = options || {};
-
   return function({ response }: Emitted) {
-    const parts: Array<string> = [];
-    if (prefix !== "") {
-      parts.push(prefix, delimiter);
-    }
-    parts.push(response.title);
-    if (suffix !== "") {
-      parts.push(delimiter, suffix);
-    }
-    document.title = parts.join(" ");
+    const title = callback(response);
+    document.title = title;
   };
 }